refactor : 컨트롤러의 상위 컨트롤러와 하위 컨트롤러 계층 구분
