+++ src/main/java/Application.java
@@ -1,5 +1,5 @@
 import config.AppConfig;
-import controller.MainController;
+import controller.facade.MainController;
 
 public class Application {
     public static void main(String[] args) {
