--- src/main/java/controller/Controller.java
@@ -1,5 +1,0 @@
-package controller;
-
-public interface Controller {
-    void run();
-}
