+++ src/main/java/controller/sub/StoreController.java
@@ -1,8 +1,8 @@
-package controller;
+package controller.sub;
 
 import service.AttendanceStoreService;
 
-public class StoreController implements Controller {
+public class StoreController implements SubController {
     private final AttendanceStoreService attendanceStoreService;
 
     public StoreController(AttendanceStoreService attendanceStoreService) {
