+++ src/main/java/controller/sub/SubController.java
@@ -1,0 +1,5 @@
+package controller.sub;
+
+public interface SubController {
+    void run();
+}
